requirements.txt
antropy==0.1.9
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
executing==2.2.0
flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
idna==3.10
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.24
langchain-core==0.3.56
lazy_loader==0.4
lightgbm==4.5.0
llvmlite==0.44.0
lspopt==1.4.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mne==1.9.0
numba==0.61.0
numpy==1.26.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pennsieve==7.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sleepecg==0.5.9
stack-data==0.6.3
tensorpac==0.6.5
threadpoolctl==3.5.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
widgetsnbextension==4.0.13
boto3==1.34.7
